requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.34
botocore==1.40.34
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
fastapi==0.116.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.11.0
jmespath==1.0.1
mangum==0.19.0
openai==1.108.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tinydb==4.8.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
